web/exmr/requirements.txt
astroid==1.6.3
autopep8==1.3.5
awsebcli==3.12.4
bitcoin==1.1.42
blessed==1.14.2
botocore==1.10.9
cement==2.8.2
certifi==2018.1.18
chardet==3.0.4
colorama==0.3.7
diff-match-patch==20121119
dj-database-url==0.5.0
Django==2.0.2
django-ckeditor==5.4.0
django-countries==5.1.1
django-import-export==1.0.0
django-js-asset==1.0.0
django-rosetta==0.8.1
django-social-share==1.1.2
django-timezone-field==2.0
django-widget-tweaks==1.4.1
dockerpty==0.4.1
docopt==0.6.2
docutils==0.14
et-xmlfile==1.0.1
gunicorn==19.7.1
idna==2.6
isort==4.3.4
jdcal==1.4
jmespath==0.9.3
lazy-object-proxy==1.3.1
mccabe==0.6.1
microsofttranslator==0.8
odfpy==1.3.6
openpyxl==2.5.3
pathspec==0.5.5
Pillow==5.0.0
polib==1.1.0
psycopg2==2.7.4
psycopg2-binary==2.7.4
pycodestyle==2.4.0
pylint==1.8.4
pyotp==2.2.6
python-bitcoinrpc==1.0
python-dateutil==2.7.2
pytz==2018.4
PyYAML==3.12
requests==2.9.1
semantic-version==2.5.0
six==1.11.0
sorl-thumbnail==12.4.1
tablib==0.12.1
tabulate==0.7.5
termcolor==1.1.0
unicodecsv==0.14.1
urllib3==1.22
wcwidth==0.1.7
websocket-client==0.47.0
wrapt==1.10.11
xlrd==1.1.0
xlwt==1.3.0
